internal/tsnet: soft-fork from tailscale.com/tsnet
The tsnet package is a user-friendly wrapper around a lot of other
tailscale packages, and hides a lot of details. Some of these would be
useful to expose to httpts.

In particular, I'd like to advertise a tag from tsnet. There is a PR
open to do this, tailscale/tailscale#8517, but there are a few things
that may be worth playing with here and soft internal/ fork is an easy
way to do that.